requirements.txt
uvicorn~=0.35.0
fastapi~=0.116.1
python-dotenv~=1.1.1